refactor(ui): extract styled span layout
